Example page for tests
The Itaú's credit card bill page is naturally very large and complicated. We
want to extract some data from it into an OFX file. To make our work easy to
test and examine, we created a simpler HTML file with only the compontents we
want to explore.

Since we will try to rely as much as possible in functions such as
document.getElementsByClassName(), we will not depend on the whole of the page
structure. We can have only the elements we want (and its subelements) most,
malybe all the time.