Use some past year to avoid unsuspected reliance on the current one.
If we use 2024 (the year when we wrote this line) as the value to assert on, it
may result on some functions relying on the current year as the value. However,
we want to use the due date year as the year for the values. By getting an
arbitrary, older year for the due date, we hope to prevent unnoticed
dependencies in the current year in our doctest asserts (and implementations).